CNTK v2 library: API changes
1) input_variable -> input
2) placeholder_variable -> forward_declaration
3) Function.replace_placeholder(s) -> Function.bind_forward_declaration(s)
4) output_variable -> UserFunction.output
5) ops.input only contains the batch axis; sequence.input contains both
the batch and sequence axes

Currently none of the examples have been modified to use the new APIs just
as a means of ensuring that the old deprecated APIs still work after being
redirected to the new APIs. Examples/tests will be changed to use the new
API in a latter commit.